Agrego useEffect para renderizado de componentes de Pagos
